pocs/miniapp/tutorial/requirements.txt
python_telegram_bot~=20.0a4
python-dotenv~=0.21.0